% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_Dup.R
\name{allDup}
\alias{allDup}
\title{Find duplicated elements}
\usage{
allDup(values)
}
\arguments{
\item{values}{any object that can be an argument of \code{duplicated()}}
}
\value{
Logical vector with TRUE for the duplicated elements
}
\description{
Finds duplicated elements without leaving the first occurrency out as
\code{duplicated()} does. *This function is not called at all*
}
