% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_pathway_matrix.R
\name{gene_pathway_matrix}
\alias{gene_pathway_matrix}
\title{Calculate membership matrix from pathway list}
\usage{
gene_pathway_matrix(alist)
}
\arguments{
\item{alist}{\code{list()} with one element for each pathway, named with the
corresponding pathway name. Each element on the list is a vector of
genes. Only the names of the elements in the vectors will be used (the
gene ID/names)}
}
\value{
Returns a numeric \code{data.frame} with 1's indicating membership
    of the gene to the pathway and 0 the lack thereof. Columns are pathways,
    rows are genes
}
\description{
From a \code{list()} of N pathways (with names corresponding to pathway
names/IDs) in which each element is a vector of gene names/IDs, creates an
NxM matrix (\code{data.frame}, actually) where N is the number of pathways
and M is the "uniqued" number of genes. The values are 1 or 0 designating
the presence of the gene in the patwhay
}
