% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_signif_matrix.R
\name{intersect_signif_matrix}
\alias{intersect_signif_matrix}
\title{Network from enriched pathways}
\usage{
intersect_signif_matrix(alist)
}
\arguments{
\item{alist}{\code{list()} with one element for each pathway. Each element
is a binary named vector, where the names are the gene names belonging
to the pathway and the values represent nominal significance (or lack
thereof)}
}
\value{
Returns a (square) \code{data.frame} with the p-values of the
    pair-wise intersections using a hypergeometric test
}
\description{
Creates a netowork from a list of enriched pathways, evaluates the
significance of all non-empty intersections between two pathways
}
