% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameGO.R
\name{nameGO}
\alias{nameGO}
\title{Fetches GO pathway names with the GO IDs}
\usage{
nameGO(clusters)
}
\arguments{
\item{clusters}{object with an \code{items} slot which contains the GO IDs
(e.g. "GO:0000009")}
}
\value{
Returns a \code{data.frame} GO names and titles
}
\description{
Fetches GO pathway names with the GO IDs
}
