% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_input.R
\name{parse_input}
\alias{parse_input}
\title{Parse the main input to the run_pathway-vis function}
\usage{
parse_input(obj, subsetsize = length(obj[[1]]), alpha = 0.05,
  colPval = "pvalue", colGeneID = "gene_id")
}
\arguments{
\item{obj}{list with two elements. The first element is a named list of
data.frames, with the names corresponding to pathway names. Each
of the data.frames contains the genes of the pathway (rows) and
p-values, ids, and log-fold changes for each of them (columns). The
second element of the object is a named list of p-values, where the
names correspond to the pathway names, and the p-values to the
significance of the enrichment of the pathway}

\item{subsetsize}{number of top pathways that will be processed. Normally,
only the top are interesting anyway, and saves a lot of time (default
all of them)}

\item{alpha}{level of significance to threshold gene expression differences
(default 0.05)}

\item{colPval}{column name in the data.frames from the second slot of the
\code{obj} that contains the p-values (column name or position)}

\item{colGeneID}{column name in the data.frames from the second slot of the
\code{obj} that contains the gene ID/names (column name of position)}
}
\value{
Returns a \code{list} with each entry being a named vector. The name
    of the element in the list corresponds to the pathway name and the
    values of the vector are the gene ID/names
}
\description{
Parses the input object to reduce the total number of pathway to analyse and
to discretise significance levels to binary (either significant or not) to
simplify downstream analysis
}
