% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_size_dist.R
\name{plot_size_dist}
\alias{plot_size_dist}
\title{Plot distribution of pathway sizes}
\usage{
plot_size_dist(alist, lowThreshold = 50, highThreshold = 1000,
  returnData = FALSE)
}
\arguments{
\item{alist}{\code{list()} with one element for each pathway, named with the
corresponding pathway name. Each element on the list is a vector of gene
names/IDs ('\code{character})}

\item{lowThreshold}{minimum number of genes in a pathway to pass the filter
(default=50)}

\item{highThreshold}{maximum number of genes in a pathway to pass the filter
(default=1000)}

\item{returnData}{if TRUE, returns the data.frame instead of the plot
(default=FALSE)}
}
\value{
ggplot with the  distribution of pathway sizes
}
\description{
Plots (using ggplot2) the distribution of pathway sizes (or returns the
data.frame that makes it)
}
