% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_map.R
\name{print_map}
\alias{print_map}
\title{Wrapper for plotting treemap of the clusters of pathways}
\usage{
print_map(dtf, index = c("title", "items"), vSize = "abslog10",
  type = "categorical", vColor = "title", title = "",
  inflate.labels = FALSE, lowerbound.cex.labels = 0,
  bg.labels = "#CCCCCCAA", position.legend = "none",
  palette = "PiYG", fontcolor.labels = c("black", "white"),
  border.col = c("black", "white"), border.lwds = c(5, 2), ...)
}
\arguments{
\item{dtf}{\code{data.frame()} with the following columns: (1) "items" -
pathway names; (2) "title" - name of the cluster (representative
pathway); (3) "abslog10" - -log10(pvalue) for the pathway (required)}

\item{index}{names of the columns for the hierarchical mapping}

\item{...}{parameters passed to \code{treeplot()} function}
}
\value{
None (only plots)
}
\description{
Plots the treemap of the clusters of pathways
}
