% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_result.R
\name{process_result}
\alias{process_result}
\title{Extract cluster titles}
\usage{
process_result(alist, amatrix, pvalues)
}
\arguments{
\item{alist}{\code{list()} with one element for each pathway, named with the
corresponding pathway name. Each element on the list is a vector of
genes. Only the names of the elements in the vectors will be used (the
gene ID/names)}

\item{amatrix}{cluster membership matrix (actually, a \code{data.frame}).
1's mean that the pathways belong to the same cluster, 0's if not. The
diagonal of the matrix is 1 if the pathway will be the title of the
cluster, -1 if it was discarded as the title}

\item{pvalues}{named numeric vector with the pathway-level p-values. The
names correspond to pathway names}
}
\value{
\code{data.frame()} with the following columns: (1) "items" -
    pathway names; (2) "title" - name of the cluster (representative
    pathway); (3) "abslog10" - -log10(pvalue) for the pathway
}
\description{
Gathers titles for the pathway clusters
}
