% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redundant_filter.R
\name{redundant_filter}
\alias{redundant_filter}
\title{Remove redundant pathways}
\usage{
redundant_filter(alist, threshold, pvalues, lowThreshold = 50,
  highThreshold = 1000)
}
\arguments{
\item{alist}{\code{list()} with one element for each pathway, named with the
corresponding pathway name. Each element on the list is a vector of
genes. Only the names of the elements in the vectors will be used (the
gene ID/names)}

\item{threshold}{similarity threshold (kappa score). Pairs of pathways (or
clusters of pathways) above this threshold will be iteratively merged
until the clusters cannot be further merged. The higher the threshold,
the less the pathways will be clustered}

\item{pvalues}{named numeric vector with the pathway-level p-values. The
names correspond to pathway names}

\item{lowThreshold}{minimum pathway size (default=50)}

\item{highThreshold}{maximum pathway size (default=1000)}
}
\value{
\code{list()} with two components, "clusterMat" with the cluster
    assignments and "simMat" with the final similarity matrix
}
\description{
Pathways are iteratively removed based on a set of criteria until there are
no pairs with a similarity above the given threshold
}
