% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rekey_go.R
\name{rekey_go}
\alias{rekey_go}
\title{Rename clusters with GO names}
\usage{
rekey_go(clusters, names)
}
\arguments{
\item{clusters}{\code{data.frame()} with the following columns: (1) "items"
- pathway names; (2) "title" - name of the cluster (representative
pathway); (3) "abslog10" - -log10(pvalue) for the pathway}

\item{names}{data.frame with the mapping between GO IDs and GO Names}
}
\value{
\code{list()} with two components, "clusterMat" with the cluster
}
\description{
Renames clusters from GO ID to GO Names. *This function is not called at
all*
}
