% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pathway_vis.R
\name{run_pathway_vis}
\alias{run_pathway_vis}
\title{Wrapper function to remove pathway redundancy}
\usage{
run_pathway_vis(input, subsetsize = length(input$geneset),
  threshold = 0.4, lowThreshold = 50, highThreshold = 1000,
  plot = TRUE, outPDF = "./treemap.pdf", plotTitle = "Treemap")
}
\arguments{
\item{input}{list with two elements. The first element is a named list of
data.frames, with the names corresponding to pathway names. Each of the
data.frames contains the genes of the pathway (rows) and p-values, ids,
and log-fold changes for each of them (columns). The second element of
the object is a named list of p-values, where the names correspond to
the pathway names, and the p-values to the significance of the
enrichment of the pathway (required)}

\item{subsetsize}{number of top pathways that will be processed. Normally,
only the top are interesting anyway, and saves a lot of time (default
all of them)}

\item{threshold}{similarity threshold (kappa score). Pairs of pathways (or
clusters of pathways) above this threshold will be iteratively merged
until the clusters cannot be further merged. The higher the threshold,
the less the pathways will be clustered (default=0.4)}

\item{lowThreshold}{minimum pathway size (default=50)}

\item{highThreshold}{maximum pathway size (default=1000)}

\item{plot}{whether to plot a figure to a PDF file (default=TRUE)}

\item{outPDF}{directory for output PDF figure (default="./treemap.pdf")}

\item{plotTitle}{title for the plot (default="Treemap")}
}
\value{
matrix with the clusters of "aggregated" pathways
}
\description{
Given an input structure with pathway- and gene-level information, remove
the redundant pathways (and plot if desired)
}
