% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_sim_mat.R
\name{update_simmat}
\alias{update_simmat}
\title{Update similarity matrix}
\usage{
update_simmat(failed, passed, simmat, alist)
}
\arguments{
\item{failed}{name of the pathway that will be discarded (character)}

\item{passed}{name of the pathway that will prevail (character)}

\item{simmat}{similarity matrix between pairs of pathways}

\item{alist}{named list of pathways with the hard-thresholded significance at the gene-level}
}
\value{
Returns an updated similarity matrix
}
\description{
Updates similarity matrix keeping a member of a pathway pair and discarding
the other
}
